Shift lerp to a utils module.
